Support nested toc on /documentation page
The /documentation page lists the toplevel links under each chapter.
This change handles the case where a toplevel link is for a nested
section by traversing the tree until a url is found.

For example:

Chapter 1:
- a: url1
- b:
  - b1:
    -bb2:
      -bbb1: url2
      -bbb2: url3
    -bb3: url4
  - b2: url5
- c: url6

The /documentation page will show this

Chapter 1
 a
 b
 c

The link for section b will go to bbb1 (url2)