Link to content from non-leaf nodes in the toc
Prior to this change, only leaf nodes could link to content.  Entries in
the table of contents (toc) with children couldn't point to content.  This
resulted in janky navigation with deep nesting.

If an entry's first child is named 'toc', the entry itself will point to the
URL for toc.  toc will not appear in the actual table of contents.  For
example:

    - OS User Guide:
        - 'toc': 'os/os_user_guide.md'
        - OS Core:
            - 'toc': 'os/os_core.md'
            - OS Fundamentals:
                - 'toc': 'os/os_fundamentals.md'
                - Overview: 'os/core_os/mynewt_os.md'

The content for OS User Guide will be os/os_user_guide.md.  The content
for OS Core will be os/os_core.md.