nimble_setup: rename my_app1 to my_proj1.
I found "my_app1" confusing as a project name.  At first I thought it
was referring to the name of the app, and that there was a mistake in
the document when a different name is later used.  I think using "proj"
rather than "app" for the project name clarifies the setup process.