Remove rogue space in stats documentation
An extra space before the closing triple-quotes for the first code block seems
to cause the clode block to not be closed and the rest of the page is messed
up. This is a difference between the Markdown renderer on Github and the Mkdocs
renderer used to create the documentation on the website.

Signed-off-by: Sam Bristow <sam.bristow@gmail.com>