mfg docs - slight rewording / rearranging
1. Move create mfg step such that it comes before the boot-build and
create-image steps.

Rationale: The mfg package only needs to be created once, whereas the
boot loader and image need to be rebuilt each time a new mfg image is
built.  It makes sense to group the steps that need to be repeated
together.

2. Remove redundant 'newt build rb_boot' line from example

This line was accidentally duplicated.

3. Slight rewording of some text to keep vocabulary consistent (mfg
package versus manufacturing image).